requirements1.txt
fastapi
uvicorn
psutil
matplotlib
aiofiles
#sqlite3
#aioredis  # You might not be using this, but it's in the original list
websockets
google-cloud-compute
google-api-python-client
Pillow
torch
python-multipart
torchvision